refactor(ptests): Move all multi-dataset tests to test_multisets.c
Extract multisets() and the NULL buffer tests from pcgns_ctest.c into
a dedicated test_multisets.c covering:
  - All ranks contributing data (round-trip correctness)
  - Mixed NULL buffers: all but the last rank contribute data
  - Rank-0-only: only rank 0 contributes; all others pass NULL

Also fix the NULL buffer guard missing in cgp_array_multi_{write,read}_data
and cgp_particle_{coord,field}_multi_{write,read}_data, which were not
patched by the earlier PR #944 fix.
